Feature: se reestructuraron los mappers cada cosa esta donde debe estar. Tambien se agrego la validacion en el servicio, repositorio y fluentvalidation para que si un codigoQr ya existe en una reventa esta no pueda crearse.
