+++ GoPass.Application/Services/Classes/EntradaService.cs
@@ -1,5 +1,4 @@
-﻿using Azure.Core;
-using GoPass.Application.Services.Interfaces;
+﻿using GoPass.Application.Services.Interfaces;
 using GoPass.Application.Utilities.Mappers;
 using GoPass.Domain.DTOs.Request.ReventaRequestDTOs;
 using GoPass.Domain.Models;
@@ -9,9 +8,11 @@ namespace GoPass.Application.Services.Classes
 {
     public class EntradaService : GenericService<Entrada>, IEntradaService
     {
+        private readonly IEntradaRepository _entradaRepository;
+
         public EntradaService(IEntradaRepository entradaRepository) : base(entradaRepository)
         {
-            
+            _entradaRepository = entradaRepository;
         }
 
         public async Task<Entrada> PublishTicket(PublishEntradaRequestDto publishEntradaRequestDto, int userId)
@@ -23,6 +24,13 @@ namespace GoPass.Application.Services.Classes
             await _genericRepository.Create(entradaToCreate);
 
             return entradaToCreate;
+        }
+
+        public async Task<bool> VerifyQrCodeAsync(string qrCode)
+        {
+            var ticketQrCode = await _entradaRepository.VerifyQrCodeExists(qrCode);
+
+            return ticketQrCode!;
         }
     }
 }
