+++ GoPass.Application/Services/Interfaces/IEntradaService.cs
@@ -7,5 +7,6 @@ namespace GoPass.Application.Services.Interfaces
     public interface IEntradaService : IGenericService<Entrada>
     {
         Task<Entrada> PublishTicket(PublishEntradaRequestDto publishEntradaRequestDto, int userId);
+        Task<bool> VerifyQrCodeAsync(string qrCode);
     }
 }
