+++ GoPass.Infrastructure/Repositories/Classes/EntradaRepository.cs
@@ -1,6 +1,7 @@
 ﻿using GoPass.Domain.Models;
 using GoPass.Infrastructure.Data;
 using GoPass.Infrastructure.Repositories.Interfaces;
+using Microsoft.EntityFrameworkCore;
 
 namespace GoPass.Infrastructure.Repositories.Classes
 {
@@ -9,6 +10,13 @@ namespace GoPass.Infrastructure.Repositories.Classes
         public EntradaRepository(ApplicationDbContext dbContext) : base(dbContext)
         {
             
+        }
+
+        public async Task<bool> VerifyQrCodeExists(string qrCode)
+        {
+            var qrCodeExist = await _dbSet.AnyAsync(u => u.CodigoQR == qrCode);
+
+            return qrCodeExist;
         }
     }
 }
