+++ GoPass.Infrastructure/Repositories/Interfaces/IEntradaRepository.cs
@@ -5,5 +5,6 @@ namespace GoPass.Infrastructure.Repositories.Interfaces
 {
     public interface IEntradaRepository : IGenericRepository<Entrada>
     {
+        Task<bool> VerifyQrCodeExists(string qrCode);
     }
 }
